vendor-qbr-analysis/data-generator/requirements.txt
pandas>=1.5.0
pyyaml>=6.0
numpy>=1.23.0
